+++ src/mappings/shared/index.ts
@@ -1,6 +1,4 @@
 export * from './burn'
-export * from './create'
 export * from './event'
 export * from './list'
 export * from './metadata'
-export * from './send'
