+++ src/mappings/v1/index.ts
@@ -1,13 +1,15 @@
 import { Interaction, unwrapRemark } from '@kodadot1/minimark/v1'
 import { unwrap } from '../utils'
 
-import { burn as consume, createCollection, list, send } from '../shared'
-import { mintItem } from '../shared/mint'
+import { burn as consume, list } from '../shared'
 import logger, { pending } from '../utils/logger'
 import { Context, RmrkInteraction } from '../utils/types'
 import { buy } from './buy'
 import { changeIssuer } from './change'
+import { createCollection } from './create'
 import { emote } from './emote'
+import { mintItem } from './mint'
+import { send } from './send'
 
 export async function mainFrame(remark: string, context: Context): Promise<void> {
   const base = unwrap(context, (_: Context) => ({ value: remark }))
